+++ MilleniumFalcon/produtos/admin.py
@@ -7,12 +7,3 @@ class AtividadeAdmin(admin.ModelAdmin):
     list_filter = ['nome']
 admin.site.register(Atividade,AtividadeAdmin)
 
-class ProjetoAdmin(admin.ModelAdmin):
-    list_display = ['nome']
-    list_filter = ['nome']
-admin.site.register(Projeto,ProjetoAdmin)
-
-class FomentoAdmin(admin.ModelAdmin):
-    list_display = ['nome']
-    list_filter = ['nome']
-admin.site.register(Fomento,FomentoAdmin)
