+++ MilleniumFalcon/projetos/admin.py
@@ -1,0 +1,11 @@
+from django.contrib import admin
+
+class ProjetoAdmin(admin.ModelAdmin):
+    list_display = ['nome']
+    list_filter = ['nome']
+admin.site.register(Projeto,ProjetoAdmin)
+
+class FomentoAdmin(admin.ModelAdmin):
+    list_display = ['nome']
+    list_filter = ['nome']
+admin.site.register(Fomento,FomentoAdmin)
